backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.0.0  # Versión específica
python-multipart==0.0.6
pytest==7.4.4
httpx  # Necesario para OpenAI